scripts/sonyc/requirements.txt
boto3==1.9.80
botocore==1.12.80
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
idna==2.8
jmespath==0.9.3
numpy==1.16.0
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
